requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.6
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environ==1.0
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
